Move ActivityCountries to components
